refactor(history-capture): split repository Git parsing
